refactor(scatter): extract axis resolution into axis_resolve.go (#331)
